requirements.txt
# Core dependencies
pydantic>=2.0.0
jinja2>=3.1.0
pyyaml>=6.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
cryptography>=41.0.0  # For credential encryption

# CLI and utilities
click>=8.1.0
rich>=13.0.0

# Content processing
PyPDF2>=3.0.0  # For PDF processing

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Email and OAuth dependencies
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
aiohttp>=3.9.0  # For OAuth redirect server

# Optional dependencies for future features
# httpx>=0.24.0  # For URL fetching
# beautifulsoup4>=4.12.0  # For HTML parsing